backend/pyproject.toml
[tool.poetry]
name = "metalyzr-backend"
version = "0.1.0"
description = "Metalyzr backend: FastAPI, Scrapy, PostgreSQL, Redis"
authors = ["Metalyzr Team <contact@metalyzr.gg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
requests = "^2.32.3"
# pandas = "^2.2.2"
# scikit-learn = "^1.5.0"
uvloop = "^0.19.0"
httpx = "^0.27.0"
gql = {extras = ["aiohttp"], version = "^3.5.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"

[tool.poetry.extras]
sqlite = ["pysqlite-binary"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.4"
pre-commit = "^3.7.0"
httpx = "^0.27.0" # For testing FastAPI

[tool.poetry.group.collectors.dependencies]
lxml = ">=4.9.0"
selenium = ">=4.15.0"
asyncio-throttle = ">=1.0.2"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 